utils-ros@0.1.5: ros, avoid blocking in logging
The Logging stuff in rosnodejs seems broken in more than one way. Somewherein the clear-throttle-log functions there seems to be something blocking that only manifests in our case when we have a failing ros topic timeout in webrtc-video. Sidestepping this now using the provided option, but we should clean this up in rosnodejs.